lz4: Update to version 1.9.4

+++ lib/lz4/lz4_compress.c
@@ -1068,7 +1068,8 @@ int LZ4_saveDict(LZ4_stream_t *LZ4_dict, char *safeBuffer, int dictSize)
 		const BYTE *const previousDictEnd =
 			dict->dictionary + dict->dictSize;
 		assert(dict->dictionary);
-		LZ4_memmove(safeBuffer, previousDictEnd - dictSize, dictSize);
+		LZ4_memmove(safeBuffer, previousDictEnd - dictSize,
+			    (size_t)dictSize);
 	}
 
 	dict->dictionary = (const BYTE *)safeBuffer;
